Add files via upload
This is an example of implementation of Continuous Variable QNN.

It is hybrid QNN implementation, the advantage is that non-linearity of the activation function is introduced by Non-Gaussian gates like Kerr or cubic phase and there is no issue like post-selection. However, there are other issues like making Kerr gate work and the implementation being hybrid.